factor some testing logic into MongoTesting, which can now be used by other packages
