+++ Tests/MongoDB/FindAndModify/FindAndModify.swift
@@ -1,59 +1,11 @@
-import BSONDecoding
-import BSONEncoding
 import MongoDB
-import Testing
+import MongoTesting
 
-func TestFindAndModify(_ tests:TestGroup, bootstrap:Mongo.DriverBootstrap) async
+struct FindAndModify:MongoTestBattery
 {
-    guard let tests:TestGroup = tests / "find-and-modify"
-    else
+    func run(_ tests:TestGroup, pool:Mongo.SessionPool, database:Mongo.Database) async throws
     {
-        return
-    }
-
-    await bootstrap.withTemporaryDatabase(named: "find-and-modify-tests", tests: tests)
-    {
-        (pool:Mongo.SessionPool, database:Mongo.Database) in
-
         let session:Mongo.Session = try await .init(from: pool)
-
-        struct Ruler:Equatable, Hashable, BSONDocumentDecodable, BSONDocumentEncodable
-        {
-            let id:String
-            let party:String
-            let since:Int
-
-            init(id:String,
-                party:String,
-                since:Int)
-            {
-                self.id = id
-                self.party = party
-                self.since = since
-            }
-
-            enum CodingKeys:String
-            {
-                case id = "_id"
-                case party
-                case since
-            }
-
-            init(bson:BSON.DocumentDecoder<CodingKeys, some RandomAccessCollection<UInt8>>)
-                throws
-            {
-                self.init(id: try bson[.id].decode(),
-                    party: try bson[.party].decode(),
-                    since: try bson[.since].decode())
-            }
-
-            func encode(to bson:inout BSON.DocumentEncoder<CodingKeys>)
-            {
-                bson[.id] = self.id
-                bson[.party] = self.party
-                bson[.since] = self.since
-            }
-        }
 
         let rulers:Mongo.Collection = "rulers"
         let states:([Ruler], [Ruler], [Ruler], [Ruler])
