+++ Tests/MongoDB/Insert/Insert.swift
@@ -1,18 +1,10 @@
 import MongoDB
-import Testing
+import MongoTesting
 
-func TestInsert(_ tests:TestGroup, bootstrap:Mongo.DriverBootstrap) async
+struct Insert:MongoTestBattery
 {
-    guard let tests:TestGroup = tests / "insert"
-    else
+    func run(_ tests:TestGroup, pool:Mongo.SessionPool, database:Mongo.Database) async throws
     {
-        return
-    }
-
-    await bootstrap.withTemporaryDatabase(named: "insert-tests", tests: tests)
-    {
-        (pool:Mongo.SessionPool, database:Mongo.Database) in
-
         let collection:Mongo.Collection = "ordinals"
         let session:Mongo.Session = try await .init(from: pool)
         do
